% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grangesUtils.R
\name{DNAstringset2gr}
\alias{DNAstringset2gr}
\title{Convert DNAStringSet to GenomicRanges}
\usage{
DNAstringset2gr(DNAss, defaultStrand = "*")
}
\arguments{
\item{DNAss}{A DNAStringSet}

\item{defaultStrand}{One of "+", "-" or "*" for the strand from which the sequences come from (default="*")}
}
\value{
A GenomicRanges object
}
\description{
Convert DNAStringSet to GenomicRanges
}
\examples{
DNAstringset2gr(Biostrings::DNAStringSet(x=c(chr1="agctgtagct",chr2="agagagagttt"))
}
