% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nanoSingleMolecule.R
\name{findGenomeMotifs}
\alias{findGenomeMotifs}
\title{Find non overlapping CG, GC and GCG/CGC motifs in genome}
\usage{
findGenomeMotifs(genome)
}
\arguments{
\item{genomeFile}{A DNAstringSet, path to FASTA file or BSgenome for the genome sequence.}
}
\value{
A GenomicRanges containing non ovelapping CG GC and GCGorCGC motifs. The context is shown in the metadata
}
\description{
Find non overlapping CG, GC and GCG/CGC motifs in genome
}
\examples{
findGenomeMotifs(ttTi5605dna)
}
