% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dSMF.R
\name{findNonOverlappingMotifs}
\alias{findNonOverlappingMotifs}
\title{Finds all CG, GC and long GCGC stretches in DNAStringSet}
\usage{
findNonOverlappingMotifs(DNAss)
}
\arguments{
\item{DNAss}{A DNAStringSet object}
}
\value{
A list of three genomic ranges for isolated CGs, isolated GCs, and CG/GC/GCG/CGC from longer runs
}
\description{
Finds all CG, GC and long GCGC stretches in DNAStringSet
}
\examples{
findNonOverlappingMotifs(DNAss)
}
