% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nanoSingleMolecule.R
\name{getIsolatedSites}
\alias{getIsolatedSites}
\title{Adds unique sites from first GRanges that are not in the second}
\usage{
getIsolatedSites(gr1, gr2)
}
\arguments{
\item{gr1}{A GenomicRanges object}

\item{gr2}{A GenomicRanges object}
}
\value{
A GenomicRanges object containing all of gr1 ranges that do not overlap with gr2
}
\description{
Adds unique sites from first GRanges that are not in the second
}
\examples{
getIsolatedSites(gr1,gr2)
}
