% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nanoSingleMolecule.R
\name{getOverlappingSites}
\alias{getOverlappingSites}
\title{Adds overlapping sites from first GRanges to the second}
\usage{
getOverlappingSites(gr1, gr2)
}
\arguments{
\item{gr1}{A GenomicRanges object}

\item{gr2}{A GenomicRanges object}
}
\value{
A GenomicRanges object containing all of gr2, and any GRanges from gr1 that overlap with gr1
}
\description{
Adds overlapping sites from first GRanges to the second
}
\examples{
getOverlappingSites(gr1,gr2)
}
