% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MCviPI_GpC}
\alias{MCviPI_GpC}
\title{Example of calling GpC methylation on GpC methylated amplicon}
\format{An object of type data.frame}
\source{
20171027_Minion_TMP_Meth
}
\usage{
data(MCviPI_GpC)
}
\description{
a PCR amplicon of ttTi5605 was GpC methylated in vitro
with MCviPI enzyme and then GpC methylation was called on
the data using nanopolish v0.11.0
}
\keyword{datasets}
