% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nanoSingleMolecule.R
\name{mergeCGandGCmat}
\alias{mergeCGandGCmat}
\title{Merge CG and GC methylation matrices}
\usage{
mergeCGandGCmat(tsvCG, tsvGC, genome)
}
\arguments{
\item{genome}{A BSgenome, DNAstring or path to fasta file for the genome of interest}

\item{CGmat}{A tab serparated values text file where individual CG motifs have been split}

\item{GCmat}{A tab serparated values text file where individual GC motifs have been split}
}
\value{
A tsv for both motifs with three contexts: HCG, GCH, GCGorCGC
}
\description{
Merge CG and GC methylation matrices
}
\examples{
mergeCGandGCtsv(splitMotifs(MSssI_CpG,"CG"),splitMotifs(MSssI_GpC,"GC"),ttTi5605dna)
}
