% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grangesUtils.R
\name{mIdxToGR}
\alias{mIdxToGR}
\title{Convert MIndex object to Genomic Ranges object}
\usage{
mIdxToGR(mIdx)
}
\arguments{
\item{mIdx}{A MIndex object}
}
\value{
A GenomicRanges object
}
\description{
Convert MIndex object to Genomic Ranges object
}
\examples{
mIdxToGR(Biostrings::vmatchPattern("ATTTAGGGTTTTAGAATACTGCCATTAATTAAAAAT",ttTi5605dna))
}
