% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MSssI_freqCmG}
\alias{MSssI_freqCmG}
\title{Example of calling CpG methylation on CpG methylated amplicon}
\format{An object of type data.frame}
\source{
20171027_Minion_TMP_Meth
}
\usage{
data(MSssI_freqCmG)
}
\description{
a PCR amplicon of ttTi5605 was CpG methylated in vitro
with MSssI enzyme and then CpG methylation was called on
the data using nanopolish v0.11.0 to call freq of methylation per site.
}
\keyword{datasets}
