% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MSssI_freqGCm}
\alias{MSssI_freqGCm}
\title{Example of calling GpC methylation on CpG methylated amplicon}
\format{An object of type data.frame}
\source{
20171027_Minion_TMP_Meth
}
\usage{
data(MSssI_freqGCm)
}
\description{
a PCR amplicon of ttTi5605 was CpG methylated in vitro
with MSssI enzyme and then GpC methylation was called on
the data using nanopolish v0.11.0 to call freq of methylation per site.
}
\keyword{datasets}
