% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dSMF.R
\name{plotSingleMoleculesAmp}
\alias{plotSingleMoleculesAmp}
\title{Do a single molecule plot of dSMF data}
\usage{
plotSingleMoleculesAmp(mat, regionName, regionGRs, featureGRs = c(),
  myXlab = "CpG/GpC position", featureLabel = "TSS", title = NULL,
  baseFontSize = 12)
}
\arguments{
\item{mat}{A methylation matrix}

\item{regionName}{The name of the region to be plotted (usually gene name). Must match Ids in reagionGRs}

\item{regionGRs}{Genomic Ranges which include region to be plotted (ID metadata column must match regionName)}

\item{featureGRs}{Genomic Ranges of any features of interest (e.g. TSS) that you wish to be plotted as vertical red line}

\item{myXlab}{Text to be used for X axis label (default="CpG/GpC position")}

\item{featureLabel}{Text to be used to label special feature you are plotting (default="TSS")}

\item{title}{Text to be used as plot title (default=NULL will results in the plot named accorind to the regionGR)}

\item{baseFontSize}{The base font size to be used for the plot (default is 12)}
}
\value{
A ggplot2 plotting object
}
\description{
Do a single molecule plot of dSMF data
}
\examples{
plotSingleMoleculesAmp(methMat,"ttTi5605",ttTi5605gr,featureGRs=c(),myXlab="CpG position",featureLabel="TSS",title=NULL,baseFontSize=12)
}
