% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dSMF.R
\name{splitGCGCruns}
\alias{splitGCGCruns}
\title{Splits GenomicRange into multiple ranges of width 2 or 3.}
\usage{
splitGCGCruns(gr)
}
\arguments{
\item{gr}{A GenomicRanges object}
}
\value{
A GenomicRanges object with multiple non-overlapping ranges width 2 or 3.
}
\description{
Splits GenomicRange into multiple ranges of width 2 or 3.
}
\examples{
splitGCGCruns(gr1)
}
