% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nanoSingleMolecule.R
\name{splitMotifs}
\alias{splitMotifs}
\title{Split sequences with multiple motifs}
\usage{
splitMotifs(tsv, motif)
}
\arguments{
\item{tsv}{A tab serparated values text file with results from Nanopore CpG or GpC methylation calls}

\item{motif}{The motif to search for: "CG" or "GC".}
}
\value{
A tsv with all mutli-motif sequences split into individual motifs with normalised log_lik_ratio
}
\description{
Split sequences with multiple motifs
}
\examples{
splitMotifs(MSssI_CpG,"CG")
}
