% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dSMF.R
\name{splitRangeInPairs}
\alias{splitRangeInPairs}
\title{Splits GenomicRange into multiple ranges of width 2.}
\usage{
splitRangeInPairs(gr)
}
\arguments{
\item{gr}{A GenomicRanges object}
}
\value{
A GenomicRanges object with multiple non-overlapping ranges width 2.
}
\description{
Splits GenomicRange into multiple ranges of width 2.
}
\examples{
splitRangeInPairs(gr1)
}
