% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nanoSingleMolecule.R
\name{tsvToGR}
\alias{tsvToGR}
\title{Convert nanopore tsv to genomic ranges}
\usage{
tsvToGR(tsv, keepCols = c("read_name", "log_lik_ratio"))
}
\arguments{
\item{tsv}{A tab serparated values text file where individual motifs have been split}

\item{keepCols}{Columns from the tsv to put in as metadata in the new GR object}
}
\value{
A genomic ranges object of motifs in tsv with some of the tsv data columns as metadata
}
\description{
Convert nanopore tsv to genomic ranges
}
\examples{
tsvToGR(splitMotifs(MSssI_CpG,"CG"))
}
