% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ttTi5605dna}
\alias{ttTi5605dna}
\title{DNAstringset for sequence used in example data}
\format{An object of type DNAStringset}
\usage{
data(ttTi5605dna)
}
\description{
DNAstringset object with fasta sequence of genome
}
\keyword{datasets}
