% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ttTi5605gr}
\alias{ttTi5605gr}
\title{GenomicRanges for sequence used in example data}
\format{An object of type GRanges}
\usage{
data(ttTi5605gr)
}
\description{
GRanges or GRangesList type object for genome
}
\keyword{datasets}
